IR: Split abstract base classes into separate source file
