+++ loki/ir/__init__.py
@@ -9,6 +9,7 @@
 The Loki internal representation (IR) and associated APIs for tree traversal.
 """
 
+from loki.ir.abstract_nodes import *  # noqaOA
 from loki.ir.expr_visitors import *  # noqa
 from loki.ir.find import *  # noqa
 from loki.ir.ir_graph import *  # noqa
